requirements.txt
absl-py==2.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Cython==3.0.8
Deprecated==1.2.14
filelock==3.13.1
fsspec==2024.2.0
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==2.8.8
numpy==1.24.4
ortools==9.8.3296
pandas==2.0.3
protobuf==4.25.2
pyconcorde @ git+https://github.com/jvkersch/pyconcorde@8a6b193b79ebdf8f07e0b0635722b3b4edbc1560
python-tsp==0.4.0
pytz==2024.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
sympy==1.12
tabulate==0.8.10
threadpoolctl==3.2.0
torch==1.9.0
tqdm==4.66.1
tsplib95==0.7.1
tzdata==2023.4
urllib3==2.2.0
wrapt==1.16.0
